# Issue# 1482: slow compile times. Quick (and arbitrary changes) to syntax makes compiling faster