TabTrayController: 60s to 7s
BrowserTable: 4s to <0.5s
SQLiteLogins: 7s to <0.5s

I didn't put too much time into the exact causes here; just changing to equivalent syntax produces an acceptable speedup.

See issue #1482